Rearrange the sound generating code
The wave base class is remove in favour of just using duck typing. It
used to contain easy way to produce adapters by calling methods, however
this was inextensible.

The wave adapters were move into their own file "wave_trans.py"

The Sequencer Item was renamed to be the Time Wave and was included with
the other general wave transforms. This change is reflected in the
sequencer, which uses the new Time Wave.